Reject inputs with potential infinite pointer loop.
While the generated parser can already automatically make sure that
parsing makes progress in the input stream, with the heavy use of
pointers in this parser (via `&parse-at`) it was possible to create
inputs which would trigger an infinite loop which was not automatically
detected.

This patch adds manual tracking of such pointer offsets and makes sure
that we parse from a single offset exactly once. This should not
affect good inputs.